requirements.txt

requests==2.26.0
numpy==1.21.4
pandas==1.3.5
matplotlib==3.5.1
transformers==4.31.0
torch==2.0.1
Pillow==9.5.0
sentence_transformers==2.1.0
faiss-cpu